Rename helper wallet reducers
