+++ packages/wallet/src/redux/reducers/initializedReducer.ts
